feat: add new wallet management features and improve UI components
- Introduced new buttons for device management, wallet backup, and referral code management in the Account Manager.
- Enhanced the wallet editing experience with new components for renaming and removing wallets.
- Updated the wallet options to include device management and backup functionalities.
- Removed deprecated components related to hidden wallet management.
- Improved the overall layout and responsiveness of wallet details pages.
- Added toast notifications for wallet import success and error handling.

This commit enhances user experience in managing wallets and devices, streamlining the process of wallet creation and management.

+++ packages/kit-bg/src/services/ServiceAccount/ServiceAccount.ts
@@ -2451,7 +2451,12 @@ class ServiceAccount extends ServiceBase {
 
   @backgroundMethod()
   async createHWWalletBase(params: IDBCreateHwWalletParams) {
-    const { features, passphraseState, fillingXfpByCallingSdk } = params;
+    const {
+      features,
+      passphraseState,
+      fillingXfpByCallingSdk,
+      isMockedStandardHwWallet,
+    } = params;
     if (!features) {
       throw new Error('createHWWalletBase ERROR: features is required');
     }
@@ -2469,7 +2474,7 @@ class ServiceAccount extends ServiceBase {
     });
 
     let xfp: string | undefined;
-    if (fillingXfpByCallingSdk) {
+    if (fillingXfpByCallingSdk && !isMockedStandardHwWallet) {
       xfp = await this.backgroundApi.serviceHardware.buildHwWalletXfp({
         connectId,
         deviceId,
@@ -2482,8 +2487,11 @@ class ServiceAccount extends ServiceBase {
       ...params,
       xfp,
       passphraseState: passphraseState || '',
-      getFirstEvmAddressFn: async () => {
-        const r =
+      getFirstEvmAddressFn: async (): Promise<string | null> => {
+        if (isMockedStandardHwWallet) {
+          return '';
+        }
+        const r: string | null =
           await this.backgroundApi.serviceHardware.getEvmAddressByStandardWallet(
             {
               connectId,
@@ -2805,18 +2813,24 @@ class ServiceAccount extends ServiceBase {
   }
 
   @backgroundMethod()
+  @toastIfError()
   async removeWallet({
     walletId,
     skipBackupWalletRemove,
+    isRemoveToMocked,
   }: Omit<IDBRemoveWalletParams, 'password' | 'isHardware'>) {
     if (!walletId) {
       throw new Error('walletId is required');
+    }
+    if (accountUtils.isOthersWallet({ walletId })) {
+      throw new Error('Remove non-hd and non-hw wallet is not allowed');
     }
     await this.backgroundApi.servicePassword.promptPasswordVerifyByWallet({
       walletId,
     });
     const result = await localDb.removeWallet({
       walletId,
+      isRemoveToMocked,
     });
     appEventBus.emit(EAppEventBusNames.WalletUpdate, undefined);
     await this.backgroundApi.serviceDApp.removeDappConnectionAfterWalletRemove({
